+++ plugins/providers/virtualbox/driver/base.rb
@@ -279,6 +279,9 @@ module VagrantPlugins
           # Variable to store our execution result
           r = nil
 
+          # If there is an error with VBoxManage, this gets set to true
+          errored = false
+
           retryable(:on => Vagrant::Errors::VBoxManageError, :tries => tries, :sleep => 1) do
             # Execute the command
             r = raw(*command, &block)
@@ -294,7 +297,7 @@ module VagrantPlugins
                 # This is usually indicative of a corrupted VirtualBox install.
                 raise Vagrant::Errors::VBoxManageNotFoundError
               else
-                raise Vagrant::Errors::VBoxManageError, :command => command.inspect
+                errored = true
               end
             else
               # Sometimes, VBoxManage fails but doesn't actual return a non-zero
@@ -302,9 +305,17 @@ module VagrantPlugins
               # occurred.
               if r.stderr =~ /VBoxManage: error:/
                 @logger.info("VBoxManage error text found, assuming error.")
-                raise Vagrant::Errors::VBoxManageError, :command => command.inspect
+                errored = true
               end
             end
+          end
+
+          # If there was an error running VBoxManage, show the error and the
+          # output.
+          if errored
+            raise Vagrant::Errors::VBoxManageError,
+              :command => command.inspect,
+              :stderr  => r.stderr
           end
 
           # Return the output, making sure to replace any Windows-style